chainstate-storage: Move `utxo_db` into the `internal` module
